feat: create verifiable election file before ending tally
