+++ bulletin_board/server/app/commands/process_tally_step.rb
@@ -51,9 +51,63 @@ class ProcessTallyStep < Rectify::Command
 
   def save_election!
     if response_log_entries && response_log_entries.any? { |le| le.message_identifier.type == "end_tally" }
+      create_verifiable_results
       election.tally_ended!
     else
       election.save!
     end
+  end
+
+  def create_verifiable_results
+    tar_filename, tar_filename_hash = create_tar do |tmpdir|
+      jsonl = nil
+      election.log_entries.find_each do |log_entry|
+        jsonl = next_file(tmpdir, jsonl) if new_file_needed?(log_entry)
+        write_log_entry(jsonl, log_entry)
+      end
+      close_file(jsonl)
+    end
+
+    election.verifiable_results_hash = tar_filename_hash
+    election.verifiable_results.attach(io: File.open(tar_filename), filename: verifiable_results_filename)
+    FileUtils.rm(tar_filename)
+  end
+
+  def verifiable_results_filename
+    @verifiable_results_filename ||= "election-#{election.id}.tar"
+  end
+
+  def create_tar
+    Dir.mktmpdir do |tmpdir|
+      yield tmpdir
+
+      `TMPFILE=$(mktemp) && cd #{tmpdir} && tar -cf $TMPFILE * && sha256sum $TMPFILE`.split.reverse
+    end
+  end
+
+  def new_file_needed?(log_entry)
+    log_entry.message_type.starts_with?(/(create|start|end)_/)
+  end
+
+  def next_file(tmpdir, jsonl)
+    if jsonl
+      close_file(jsonl)
+    else
+      jsonl = { status_index: 0 }
+    end
+
+    status_title = Election.statuses.to_a[jsonl[:status_index]].reverse.join("-")
+    jsonl[:file] = Zlib::GzipWriter.open(File.join(tmpdir, "#{status_title}.jsonl.gz"))
+    jsonl[:status_index] += 1
+    jsonl
+  end
+
+  def close_file(jsonl)
+    jsonl[:file].close
+  end
+
+  def write_log_entry(jsonl, log_entry)
+    jsonl[:file].write(log_entry.slice(:message_id, :signed_data, :chained_hash, :content_hash).to_json)
+    jsonl[:file].write("\n")
   end
 end
