+++ bulletin_board/server/app/commands/publish_results.rb
@@ -35,8 +35,6 @@ class PublishResults < Rectify::Command
       log_entry.election = election
       log_entry.save!
 
-      create_verifiable_results
-
       election.results_published!
     end
 
@@ -46,57 +44,4 @@ class PublishResults < Rectify::Command
   private
 
   attr_accessor :authority
-
-  def verifiable_results_filename
-    @verifiable_results_filename ||= "election-#{election.id}.tar"
-  end
-
-  def create_verifiable_results
-    tar_filename, tar_filename_hash = create_tar do |tmpdir|
-      jsonl = nil
-      election.log_entries.find_each do |log_entry|
-        jsonl = next_file(tmpdir, jsonl) if new_file_needed?(log_entry)
-        write_log_entry(jsonl, log_entry)
-      end
-      close_file(jsonl)
-    end
-
-    election.verifiable_results_hash = tar_filename_hash
-    election.verifiable_results.attach(io: File.open(tar_filename), filename: verifiable_results_filename)
-    FileUtils.rm(tar_filename)
-  end
-
-  def create_tar
-    Dir.mktmpdir do |tmpdir|
-      yield tmpdir
-
-      `TMPFILE=$(mktemp) && cd #{tmpdir} && tar -cf $TMPFILE * && sha256sum $TMPFILE`.split.reverse
-    end
-  end
-
-  def new_file_needed?(log_entry)
-    log_entry.message_type.starts_with?(/(create|start|end|publish)_/)
-  end
-
-  def next_file(tmpdir, jsonl)
-    if jsonl
-      close_file(jsonl)
-    else
-      jsonl = { status_index: 0 }
-    end
-
-    status_title = Election.statuses.to_a[jsonl[:status_index]].reverse.join("-")
-    jsonl[:file] = Zlib::GzipWriter.open(File.join(tmpdir, "#{status_title}.jsonl.gz"))
-    jsonl[:status_index] += 1
-    jsonl
-  end
-
-  def close_file(jsonl)
-    jsonl[:file].close
-  end
-
-  def write_log_entry(jsonl, log_entry)
-    jsonl[:file].write(log_entry.slice(:message_id, :signed_data, :chained_hash, :content_hash).to_json)
-    jsonl[:file].write("\n")
-  end
 end
