+++ bulletin_board/server/app/graphql/decidim_bulletin_board_schema.rb
@@ -4,7 +4,7 @@ class DecidimBulletinBoardSchema < GraphQL::Schema
   mutation(Types::MutationType)
   query(Types::QueryType)
   use GraphQL::Guard.new(
-    policy_object: Policies::VisibilityForAll,
+    policy_object: Policies::VisibilityPolicy,
     not_authorized: ->(_type, _field) { nil }
   )
 end
